+++ Sources/Mockolo/Executor.swift
@@ -206,7 +206,7 @@ class Executor {
         do {
             try generate(sourceDirs: srcDirs,
                          sourceFiles: srcs,
-                         parser: ParserViaSwiftSyntax(),
+                         parser: SourceParser(),
                          exclusionSuffixes: exclusionSuffixes,
                          mockFilePaths: mockFilePaths,
                          annotation: annotation,
@@ -232,12 +232,4 @@ class Executor {
             fatalError("Generation error: \(error)")
         }
     }
-}
-
-public struct Version {
-    /// The string value for this version.
-    public let value: String
-
-    /// The current Mockolo version.
-    public static let current = Version(value: "1.3.2")
 }
