+++ Sources/MockoloFramework/Version.swift
@@ -1,0 +1,7 @@
+public struct Version {
+    /// The string value for this version.
+    public let value: String
+
+    /// The current Mockolo version.
+    public static let current = Version(value: "1.3.2")
+}
